server/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Paul Bailey", email = "paul@neutron.studio"},
]
dependencies = [
    "django>=4.2.1",
    "psycopg2-binary>=2.9.6",
    "django-timescaledb>=0.2.13",
    "djangorestframework>=3.14.0",
    "django-filter>=23.2",
    "django-cors-headers>=4.0.0",
    "django-environ>=0.10.0",
    "pyjwt>=2.7.0",
]
requires-python = ">=3.10"
license = {text = "BSDv3"}

[tool]
[tool.pdm]
[tool.pdm.scripts]
_.env_file = ".env"
req = "pdm export -o requirements.txt --without-hashes --prod"
